Change generator state dict to hold one metadata per entry (fixes #26)
The previous concept to have one metadata object inside the state dict was not compatible with non-sequential augmentation, e.g., the separate selection of two windows.
To allow this, each key in the state dict is now associated with both a data and a metadata entry.
In addition, all default keys are now called X, with the key y envisioned for labels.
Therefore, as long as users build only sequential augmentation pipelines using the standard augmentations, they don't have to worry about keys at all.
This commit should make it easier to build complex augmentation pipelines without unexpected behaviour.
On the downside, this change makes it slightly more complicated to write custom augmentations.