Introduce blocksize attribute to writer
Writing to HDF5 files becomes very slow when writing many datasets into the same group. To mitigate this, groups can be split up internally. As datasets can be addressed similar to files in a filesystem using slashes in the string, this does not require further changes in the SeisBench library. If required, the writer automatically names traces block/trace and writes this into the metadata trace_name column. Also introduces a warning if large datasets are written without blocks.