Reduce non-annotated portion in WaveformModel annotate mode "array"
So far, _annotate_array only used as many windows as could be fully fitted while keeping the given overlap. This might lead to up to nearly a full non-annotated window at the end of a trace. The new version now adds one additional window to the trace, if necessary, which will have potentially larger overlap, but produces predictions for the remaining samples.