Fix ETHZ dataset issue with picks on different channels
Picks annotated on different channels of the same station for the same event where stored in separate traces. This lead to issues, as in some cases the channel in QuakeML was HHZ, while in others it was HH, without a component identifier. This commit treats all picks for an event with the same NET.STA.LOC together and stores them in the same trace.