Fix bug when using writer without trace_name in metadata
The WaveformDataWriter tried to use the trace_name parameter from metadata when writing out a bucket containing only a single trace. This leads to a crash if no trace name is provided. With this commit, the trace_name parameter is only used if available. Instead, the next bucket name is used. Also adds a test for this issue.