Added metadata lookup cache
So far, WaveformDataset relied on pandas iloc to access metadata. However, this comes at a performance hit, that can be substantial for large datasets. This PR adds the option for a metadata cache as a list of dicts that can be access substantially faster. On the downside, this requires storing two copies of the metadata in memory. As for most cases metadata is rather small, this should usually not be a concern. The optimisation is activated by default.