Optimize data access with redundant traces
Currently, if the same trace was accessed multiple times within on query, it was processed (loaded, resampled, ...) multiple times and several copies were returned. This happens in particular when using grouping. Instead, now the trace is loaded exactly once and references to this one copy are returned. This reduces runtime and memory footprint.