Add weight version warnings
Sometimes older weights have issues and need to be reconfigured. As by default, SeisBench does not check for updates on every call to from_pretrained, this is not automatically detected. As a mitigation, this PR allows to hardcode version warnings. This way, a new version of SeisBench will alert the users if they are using deprecated weights.