Allow AddGap and ChannelDropout to modify labels
Add an option for AddGap and ChannelDropout to modify labels directly.
For AddGap, the gap can be added to both the waveform and the labels.
For ChannelDropout, if all channels are zeros, labels will also be
zeroed out. If there is a noise label, the noise label within the gap
is 1.