Change annotate logic from windowed to batched pre- and postprocessing
Pre- and postprocessing in the annotate function was so far done per window and in numpy. This PR changes the processing to a batched mode running on PyTorch. This also ensures that all processing can happen on the GPU if available. In addition, this now allows to wrap the three processing steps in a thread, allowing to release the GIL during the heavy numeric operations. Overall, this will reduce run times, in particular, on GPU.

As this change improves the efficiency of the asyncio implementation, it removes the (suboptimal) process implementation completely.

To ensure compatibility with the previous implementation, a suite of tests has been added and the former `annotate_window_pre`/`annotate_window_post` functions are retained. These functions and tests should be removed in a future PR.

Additional small changes:
- increased default batch size from 64 to 256
- refactored some functions in seisbench.util