Add warnings for use of default component order
Warn the user when they don't specify a component order and the default value is
used. Also warn again if the default component order doesn't match what is the
dataset's data format.

Change tests that check for warnings to clear the message log after loading the
dummy dataset, not before, so that the new warnings don't break any tests.