"Inclusion of Domain-Knowledge PhaseNet (DKPN) picker as published in Lomax et al. 2024 (https://doi.org/10.26443/seismica.v3i1.1164) and the some pre-trained weights. The code has been refactored from its original repository (https://github.com/INGV/DKPN) to suit v0.8 of seisbench (originally developed over v0.3/0.4).
Everything seems fine, but probably double check needed by the mantainer(s).

The end-user is referred to the original paper for details and to the original repository for the training scripts.

Additional bug-fix for `base.WaveformModel` included that double check the right device for the machine at line 489-492 during execution of `from_pre_trained` method. This should cover cases where model-weights are trained/stored over a GPU but later used on a CPU-only machine."