Introduce `AbstractBenchmarkDataset` class
So far, `BenchmarkDataset` inherited from `WaveformDataset`. However, this meant that it was impossible to reuse the download functionality for datasets with a different file structure. To fix this, all logic from the `BenchmarkDataset` was moved to an `AbstractBenchmarkDataset`. `BenchmarkDataset` now inherits from `AbstractBenchmarkDataset` (the download logic) and from `WaveformDataset` (the logic for handling the actual waveform data). The class BenchmarkDataset is renamed to `WaveformBenchmarkDataset`. For downward compatibility the alias `BenchmarkDataset` is kept, but issues a deprecation warning when used.